open-brain-mcp/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
asyncpg>=0.29.0
httpx>=0.27.0
mcp>=1.0.0
python-dateutil>=2.9.0
pydantic>=2.0.0
